# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DocKonvert API",
    "version": "0.1.0",
    "description": "Convert any document to Markdown or PDF via a simple REST API. Powered by Docling.\n\n## Authentication\nAll API requests require an API key passed via the `X-API-Key` header.\n\n```\nX-API-Key: dk_your_api_key\n```\n\nGet your free API key by [creating an account](https://dockonvert.com/register).\n\n## Rate Limits\n- **Free tier**: 100 requests per hour\n- **Max file size**: 10MB\n\nRate limit headers are included in every response:\n```\nX-RateLimit-Limit: 100\nX-RateLimit-Remaining: 99\nX-RateLimit-Reset: 1700000000\n```\n\n## Document Retention\nDocuments are retained for 24 hours for download purposes, then automatically and permanently purged.",
    "contact": {
      "name": "BRAMKAS INC",
      "url": "https://www.bramkas.com",
      "email": "hello@bramkas.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://dockonvert.com/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8009/api",
      "description": "Local Development"
    },
    {
      "url": "https://api.dockonvert.com/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "operationId": "createJob",
        "summary": "Create a conversion job",
        "description": "Create an async job with import/convert/export pipeline. Supports URL import, 5 output formats, webhooks, and S3 export.",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["tasks"], "properties": { "tasks": { "type": "object" }, "webhook": { "type": "object" }, "tag": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Job created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } } }
      },
      "get": {
        "operationId": "listJobs",
        "summary": "List jobs",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "tag", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "List of jobs", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } } } } } }
      }
    },
    "/v1/jobs/upload": {
      "post": {
        "operationId": "createUploadJob",
        "summary": "Create job with file upload",
        "description": "Upload a file and specify output format. Auto-creates import/convert/export pipeline.",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["file", "output_format"], "properties": { "file": { "type": "string", "format": "binary" }, "output_format": { "type": "string", "enum": ["markdown", "pdf", "html", "txt", "docx"] }, "options": { "type": "string" }, "webhook_url": { "type": "string" }, "tag": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Job created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } } }
      }
    },
    "/v1/jobs/{jobId}": {
      "get": {
        "operationId": "getJob", "summary": "Get job status", "tags": ["Jobs"], "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Job details" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "operationId": "deleteJob", "summary": "Delete a job", "tags": ["Jobs"], "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/v1/jobs/{jobId}/download": {
      "get": {
        "operationId": "downloadJobResult", "summary": "Download job output", "tags": ["Jobs"], "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "task", "in": "query", "schema": { "type": "string", "default": "convert" } }
        ],
        "responses": { "200": { "description": "File download" }, "400": { "description": "Not completed" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for conversion endpoints. Get one by registering at /auth/register."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /auth/login or /auth/register."
      }
    },
    "schemas": {
...
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "processing", "completed", "failed"] },
          "tag": { "type": "string", "nullable": true },
          "tasks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "operation": { "type": "string" },
                "status": { "type": "string" },
                "result": { "type": "object", "nullable": true },
                "error": { "type": "string", "nullable": true }
              }
            }
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time", "nullable": true },
          "download_url": { "type": "string", "nullable": true }
        }
      },
...
    },
...
  }
}